code/requirements.txt
cassandra-driver
pandas